Cut spatial networks in folded tissues

Opens an interactive viewer where a spatial network is visualized on top of an H&E image.

Usage

CutSpatialNetwork(
  object,
  datadir = NULL,
  sampleID = 1,
  host = "127.0.0.1",
  port = 8080L,
  container_width = "800px",
  container_height = "800px",
  overwrite_network_json = TRUE,
  verbose = TRUE
)

Arguments

object

A Seurat object created with semla

datadir

A directory containing network data and image tiles

sampleID

An integer specifying a tissue section in the Seurat object

host

A string with a valid IPv4 or IPv6 address to listen on, Defaults to localhost "127.0.0.1"

port

An integer that indicates the port to listen on.

container_width, container_height

Set height and width of container

overwrite_network_json

Logical specifying if the JSON file containing the spatial network should be overwritten after completion

verbose

Print messages

Value

A tbl_graph object representing a spatial network

Details

Each spot is connected to adjacent neighbors by edges and the edges can be cut by holding the SHIFT key while moving the cursor across them. Cut edges can be mended by holding the CTRL key while moving the cursor across them. The aim is to cut edges between spots located in separate layers. The output is a tbl_graph object representing the spatial network which can be processed further with AdjustTissueCoordinates to perform "digital unrolling".

Examples

A tutorial can be found on our [package website](https://ludvigla.github.io/semla/). Got to tutorials -> Digital unrolling

See also

Author

Ludvig Larsson