Update an image_info tibble

Takes two tbl objects as input, one with image information produced with LoadImageInfo and one with scale factors produced with LoadScaleFactors. The scale factors are used to calculate the width and height of the original H&E image which will be added to the output tbl.

Usage

UpdateImageInfo(image_info, scalefactors)

Arguments

image_info

A tbl object with image info

scalefactors

A tbl object with scale factors

Value

A tbl object with image info

See also